Added some helper tools, fixed some paths for eos_stb dropped_events task
