+++ .gitignore
@@ -11,3 +11,5 @@
 venv*
 
 \.vscode/
+tools/config
+checkpoints
