+++ configuration/basic_analytics/eos_stb/usagecollector_dropped_events.yml
@@ -1,0 +1,25 @@
+spark:
+  appName: UsageCollectorReport_missed_events
+  master: local[*]
+  checkpointLocation: file:///spark/checkpoints/procs/
+  shuffle.partitions: 1
+  consoleWriter: false
+  outputMode: update
+
+kafka:
+  maxOffsetsPerTrigger: 100000
+  bootstrap.servers: 172.31.101.138:9094,172.31.101.139:9094,172.31.101.140:9094
+  startingOffsets: earliest
+  topics:
+    dlq: odhdev_pro_deadletterqueue
+    inputs:
+      - odhdev_test_in_bdebotte_01
+    output: odhdev_test_out_bdebotte_01
+  security.protocol: SASL_PLAINTEXT
+  sasl.mechanism: PLAIN
+  failOnDataLoss: false
+
+analytics:
+  componentName: odhdev.eosstb.be.tst.stb
+  watermark: '1 minutes'
+  window: '1 minutes'
