--- configuration/stb_usagecollector_dropped_events.yml
@@ -1,26 +1,0 @@
-spark:
-  appName: StbUsageCollectorDroppedEvents
-  master: local[*]
-  checkpointLocation: file:///spark/checkpoints/procs/
-  shuffle.partitions: 1
-  consoleWriter: true
-  outputMode: update
-
-kafka:
-  maxOffsetsPerTrigger: 100000
-  bootstrap.servers: 192.168.100.1:9094,192.168.100.2:9094,192.168.100.3:9094,192.168.100.4:9094,192.168.100.5:9094,192.168.100.6:9094,192.168.100.7:9094,192.168.100.8:9094,192.168.100.9:9094,192.168.100.10:9094,192.168.100.11:9094,192.168.100.12:9094,192.168.100.13:9094,192.168.100.14:9094,192.168.100.15:9094,192.168.100.16:9094,192.168.100.17:9094,192.168.100.18:9094
-  startingOffsets: earliest
-  topics:
-    dlq: odhecx_pro_deadletterqueue
-    inputs:
-      - odhecx_pr_eosdtv_be_tst_stb_statelog_v1
-    output: test_bdebotte_stbusagecollectordroppedevents
-    # pr_eosdtv_be_tst_stb_usage_collector_dropped_events_v1
-  security.protocol: SASL_PLAINTEXT
-  sasl.mechanism: PLAIN
-  failOnDataLoss: false
-
-analytics:
-  componentName: vagrant.eosdtv.lab5aobo.tst.heapp.thinkanalytics.reingest
-  watermark: '1 minutes'
-  window: '1 minutes'
