+++ src/common/basic_analytics/aggregations.py
@@ -6,7 +6,7 @@ Aggregation class defines common methods for all aggregations.
 import re
 from abc import ABCMeta, abstractmethod
 
-from pyspark.sql.functions import lit, avg, count, col, regexp_replace
+from pyspark.sql.functions import lit, avg, count, col, regexp_replace, sum, min, max
 from pyspark.sql.functions import window, concat_ws, approx_count_distinct
 from pyspark.sql.types import DecimalType
