+++ tools/log4j.properties
@@ -1,0 +1,10 @@
+app=spark
+# Define the types of logger and level of logging
+log4j.rootLogger = WARN,console
+
+# Define Console Appender
+log4j.appender.console=org.apache.log4j.ConsoleAppender
+
+# Define the layout for console appender. If you do not
+# define it, you will get an error
+log4j.appender.console.layout=org.apache.log4j.PatternLayout
