+++ tools/spark_jaas.conf
@@ -1,0 +1,5 @@
+KafkaClient {
+   org.apache.kafka.common.security.plain.PlainLoginModule required
+   username="spark"
+   password="Spark";
+};
